Move sklearn I/F into a separated module
